+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button } from '../Botton/Button';
+import { Button } from '../Button';
 import './header.css';
 
 export interface HeaderProps {
