+++ src/stories/Button.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
 
-import { Button, ButtonProps } from '../components/Botton/Button';
+import { Button, ButtonProps } from '../components/Button';
 
 export default {
   title: 'Example/Button',
