+++ src/stories/Header.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
 
-import { Header, HeaderProps } from '../components/Header/Header';
+import { Header, HeaderProps } from '../components/Header';
 
 export default {
   title: 'Example/Header',
