--- src/stories/Page.stories.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { Story, Meta } from '@storybook/react';
-
-import { Page, PageProps } from '../components/Page/Page';
-import * as HeaderStories from './Header.stories';
-
-export default {
-  title: 'Example/Page',
-  component: Page,
-} as Meta;
-
-const Template: Story<PageProps> = args => <Page {...args} />;
-
-export const LoggedIn = Template.bind({});
-LoggedIn.args = {
-  ...HeaderStories.LoggedIn.args,
-};
-
-export const LoggedOut = Template.bind({});
-LoggedOut.args = {
-  ...HeaderStories.LoggedOut.args,
-};
